+++ doc/tutorials/content/sources/registration_api/example2.cpp
@@ -138,13 +138,13 @@ computeTransformation (const PointCloud<PointXYZ>::Ptr &src,
                             keypoints_tgt (new PointCloud<PointXYZ>);
 
   estimateKeypoints (src, tgt, *keypoints_src, *keypoints_tgt);
-  print_info ("Found %lu and %lu keypoints for the source and target datasets.\n", keypoints_src->size (), keypoints_tgt->size ());
+  print_info ("Found %zu and %zu keypoints for the source and target datasets.\n", static_cast<std::size_t>(keypoints_src->size ()), static_cast<std::size_t>(keypoints_tgt->size ()));
 
   // Compute normals for all points keypoint
   PointCloud<Normal>::Ptr normals_src (new PointCloud<Normal>), 
                           normals_tgt (new PointCloud<Normal>);
   estimateNormals (src, tgt, *normals_src, *normals_tgt);
-  print_info ("Estimated %lu and %lu normals for the source and target datasets.\n", normals_src->size (), normals_tgt->size ());
+  print_info ("Estimated %zu and %zu normals for the source and target datasets.\n", static_cast<std::size_t>(normals_src->size ()), static_cast<std::size_t>(normals_tgt->size ()));
 
   // Compute FPFH features at each keypoint
   PointCloud<FPFHSignature33>::Ptr fpfhs_src (new PointCloud<FPFHSignature33>),
